Netcoreapp2.0move (#603)
* build vstest.console.dll as netcoreapp2.0

* test script changes

* Fixing acceptance tests - core runner frameowork as netcoreapp2.0

* stream test run to console

* Move TestPlatform.CLI artifacts to netcoreapp2.0 folder

* Ignore failing platform tests

* Install netcoreapp1.1 shared runtime.

* Match System.Xml.XPath.XmlDocument and DependencyModel version with dotnet cli

* Install netcoreapp 1.1.1 shared runtime for acceptance tests targeting netcoreapp1.1.
Use async stdout, stderr listeners for IntegrationTestBase to ensure there are no deadlocks
in the child processes created for acceptance tests.

* Revert nuspec2 files

* uncomment build steps

* Move DataCollectors to netcoreapp2.0. Use the datacollector for the runner framework
for acceptance tests.

* Enable vstest.console and translation layer unit tests.
Fix path for datacollection acceptance tests sample data collector.

* Add target fx as a parameter to run specific tests in test.ps1.
Call WaitForExit twice in IntegrationTestBase to ensure streams are flushed.
Ignore a failing test in net46 temporarily.

* Make error null to handle JSON serialize exception

* Use netcoreapp2.0 as the base directory for core clr package.

* Skip target framework only if not null.

+++ test/Microsoft.TestPlatform.AcceptanceTests/DataCollectionTests.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 namespace Microsoft.TestPlatform.AcceptanceTests
@@ -8,15 +8,15 @@ namespace Microsoft.TestPlatform.AcceptanceTests
     using System.IO;
     using System.Xml;
 
+    using Microsoft.TestPlatform.TestUtilities;
     using Microsoft.VisualStudio.TestPlatform.ObjectModel;
     using Microsoft.VisualStudio.TestPlatform.Utilities.Helpers;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
-    using Microsoft.TestPlatform.TestUtilities;
 
     [TestClass]
     public class DataCollectionTests : AcceptanceTestBase
     {
-        private string resultsDir;
+        private readonly string resultsDir;
 
         public DataCollectionTests()
         {
@@ -26,7 +26,10 @@ namespace Microsoft.TestPlatform.AcceptanceTests
         [TestCleanup]
         public void Cleanup()
         {
-            Directory.Delete(this.resultsDir, true);
+            if (Directory.Exists(this.resultsDir))
+            {
+                Directory.Delete(this.resultsDir, true);
+            }
         }
 
         [CustomDataTestMethod]
@@ -34,14 +37,14 @@ namespace Microsoft.TestPlatform.AcceptanceTests
         [NETCORETargetFramework]
         public void ExecuteTestsWithDataCollection(string runnerFramework, string targetFramework, string targetRuntime)
         {
-            AcceptanceTestBase.SetTestEnvironment(this.testEnvironment, runnerFramework, targetFramework, targetRuntime);
+            SetTestEnvironment(this.testEnvironment, runnerFramework, targetFramework, targetRuntime);
 
             var assemblyPaths = this.BuildMultipleAssemblyPath("SimpleTestProject2.dll").Trim('\"');
             string runSettings = this.GetRunsettingsFilePath();
             string diagFileName = Path.Combine(this.resultsDir, "diaglog.txt");
-
             var arguments = PrepareArguments(assemblyPaths, this.GetTestAdapterPath(), runSettings, this.FrameworkArgValue);
             arguments = string.Concat(arguments, $" /ResultsDirectory:{resultsDir}", $" /Diag:{diagFileName}");
+
             this.InvokeVsTest(arguments);
 
             this.ValidateSummaryStatus(1, 1, 1);
@@ -126,6 +129,8 @@ namespace Microsoft.TestPlatform.AcceptanceTests
 
             dataCollectionAttributes.Add("friendlyName", "SampleDataCollector");
             dataCollectionAttributes.Add("uri", "my://sample/datacollector");
+
+            // Data collector needs to be targeted to same runtime as the runner framework
             var codebase = Path.Combine(
                 this.testEnvironment.TestAssetsPath,
                 Path.GetFileNameWithoutExtension("OutOfProcDataCollector"),
@@ -133,10 +138,9 @@ namespace Microsoft.TestPlatform.AcceptanceTests
                 this.testEnvironment.BuildConfiguration,
                 this.testEnvironment.RunnerFramework,
                 "OutOfProcDataCollector.dll");
+            Assert.IsTrue(File.Exists(codebase), "Data collector assembly not found: " + codebase);
 
-            var outOfProcAssemblyPath = this.testEnvironment.GetTestAsset("OutOfProcDataCollector.dll");
-
-            dataCollectionAttributes.Add("assemblyQualifiedName", string.Format("OutOfProcDataCollector.SampleDataCollector, {0}", AssemblyUtility.GetAssemblyName(outOfProcAssemblyPath)));
+            dataCollectionAttributes.Add("assemblyQualifiedName", string.Format("OutOfProcDataCollector.SampleDataCollector, {0}", AssemblyUtility.GetAssemblyName(codebase)));
             dataCollectionAttributes.Add("codebase", codebase);
             CreateDataCollectionRunSettingsFile(runsettingsPath, dataCollectionAttributes);
             return runsettingsPath;
